wip: support remote debugging output redirection of windows
