+++ service/dap/server.go
@@ -11,8 +11,6 @@ package dap
 import (
 	"bufio"
 	"bytes"
-	"crypto/rand"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -33,7 +31,6 @@ import (
 	"strconv"
 	"strings"
 	"sync"
-	"syscall"
 	"time"
 
 	"github.com/go-delve/delve/pkg/gobuild"
@@ -1155,8 +1152,15 @@ func (s *Session) onLaunchRequest(request *dap.LaunchRequest) {
 				log.Fatal(err)
 			}
 
-			go readerFunc(stdoutFile, "stdout")
-			go readerFunc(stderrFile, "stderr")
+			go func() {
+				readerFunc(stdoutFile, "stdout")
+				os.Remove(redirects[0])
+			}()
+
+			go func() {
+				readerFunc(stderrFile, "stderr")
+				os.Remove(redirects[1])
+			}()
 		}()
 	}
 
@@ -3909,30 +3913,4 @@ func parseLogPoint(msg string) (bool, *logMessage, error) {
 		format: string(formatSlice),
 		args:   args,
 	}, nil
-}
-
-func generateStdioTempPipes() (res [2]string, err error) {
-	r := make([]byte, 4)
-	if _, err := rand.Read(r); err != nil {
-		return res, err
-	}
-
-	var (
-		prefix     = filepath.Join(os.TempDir(), hex.EncodeToString(r))
-		stdoutPath = prefix + "stdout"
-		stderrPath = prefix + "stderr"
-	)
-
-	if err := syscall.Mkfifo(stdoutPath, 0o600); err != nil {
-		return res, err
-	}
-
-	if err := syscall.Mkfifo(stderrPath, 0o600); err != nil {
-		os.Remove(stdoutPath)
-		return res, err
-	}
-
-	res[0] = stdoutPath
-	res[1] = stderrPath
-	return res, nil
 }
